[ADD] Add new hookpoint for version >=3.0.9
